(feat) avatar updates
