+++ server/routes/userRouter.js
@@ -10,8 +10,8 @@ router.post('/', function(req,res,next){
   userController.addUser(req, res, next);
 });
 
-router.post('/', function(req,res,next){
-  userController.addUser(req, res, next);
+router.post('/avatars/', function(req,res,next){
+  userController.addAvatar(req, res, next);
 });
 
 module.exports = router;
